requirements.txt
Django==4.2.27
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
django-extensions==3.2.3
drf-spectacular==0.26.5
pyotp==2.9.0
qrcode==7.4.2
pytest==7.4.3
pytest-django==4.7.0
faker==20.1.0
